pocketllm-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
asyncpg==0.29.0
supabase==2.3.4
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.8.0
structlog==24.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.4
tenacity==8.2.3
uuid6==2024.1.12
